new power set translation to abandonment to save time
Instead of generating the power sets of customers to abandoned EACH
SUBPROBLEM, I have created a method to translate from the full power set
(all customers) to each subset of customers.  Of note: this method
short-circuits when the element numbers exceed what is available (ie not
element 5 in 4 element subset, and nothing beyond this would work as
I've used binary encoding).